Rename package(feature_pokemon to feature.pokemon)
