+++ app/src/main/java/com/hanbitkang/mypokemon/navigation/MpNavHost.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.hanbitkang.feature.favorite.navigation.favoriteGraph
-import com.hanbitkang.feature_pokemon.navigation.PokemonDestination
-import com.hanbitkang.feature_pokemon.navigation.pokemonGraph
+import com.hanbitkang.feature.pokemon.navigation.PokemonDestination
+import com.hanbitkang.feature.pokemon.navigation.pokemonGraph
 
 @Composable
 fun MpNavHost(
