+++ app/src/main/java/com/hanbitkang/mypokemon/ui/MpAppState.kt
@@ -9,7 +9,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.hanbitkang.core.designsystem.icon.MpIcons
 import com.hanbitkang.feature.favorite.navigation.FavoriteDestination
-import com.hanbitkang.feature_pokemon.navigation.PokemonDestination
+import com.hanbitkang.feature.pokemon.navigation.PokemonDestination
 import com.hanbitkang.mypokemon.navigation.TopLevelDestination
 
 @Composable
@@ -41,7 +41,7 @@ class MpAppState(
             FavoriteDestination.destination,
             MpIcons.Bookmark,
             MpIcons.BookmarkBorder,
-            com.hanbitkang.feature_pokemon.R.string.pokemon_list
+            com.hanbitkang.feature.pokemon.R.string.pokemon_list
         )
     )
 }
