+++ feature-favorite/src/androidTest/java/com/hanbitkang/feature/favorite/ExampleInstrumentedTest.kt
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.hanbitkang.feature_favorite.test", appContext.packageName)
+        assertEquals("com.hanbitkang.feature.favorite.test", appContext.packageName)
     }
 }
