+++ feature-pokemon/src/main/java/com/hanbitkang/feature/pokemon/PokemonScreen.kt
@@ -1,4 +1,4 @@
-package com.hanbitkang.feature_pokemon
+package com.hanbitkang.feature.pokemon
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
