+++ feature-pokemon/src/test/java/com/hanbitkang/feature/pokemon/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.hanbitkang.feature_pokemon
+package com.hanbitkang.feature.pokemon
 
 import org.junit.Test
